requirements.txt
streamlit
websockets
pyyaml
requests
python-dotenv
